Add : Hackerrank (warmup)
Bob has a strange counter. At the first second,1 , it displays the number 3. At each subsequent second, the number displayed by the counter decrements by 1.
The counter counts down in cycles. In the second after the counter counts down to 1 , the number becomes  2x the initial number for that countdown cycle and then continues counting down from the new initial number in a new cycle.
Given a time, , find and print the value displayed by the counter at time .