Add : Codechef (Oct challenge)
You start with an empty set of points on a two-dimensional plane. There are N queries, each with one new point (xi, yi) to insert to the set. After each query you must find the diameter of the set, i.e. the maximum distance between two points in the set. To avoid printing floating numbers, after each query print the square of the diameter.

In many problems, in test generators an author should create some (possibly smart and tricky) data, and then shuffle it before printing. Limak did the same in this problem, not noticing how crucial it may be. You can assume that in every test the N points are randomly shuffled. In other words, generators have a line "random_shuffle(array with points)" just before printing. It isn't guaranteed that a problem is solvable within the time limit without this extra constraint.