Add : ICPC C 2016 online prelims
Watson is going to play mind games with Sherlock. He gives an integer d to Sherlock.

Among all possible positive integers N having sum of digits d, consider the integer N+1, and find the minimum possible digit sum N+1 can have.

For example, if d = 9, then N could be 9, 18, 27, 36,711 and so on. However, if Sherlock chooses N = 9, then sum of digits of N + 1 i.e. 10 will be 1, which is the minimum possible.