Add : Spoj problem
Given a number, output how many digits in number that number itself is divisible by. Count all occurences of such digits in the number, not just the first.

For example, 12345 is divisible by 1, 3, and 5, so, the answer in this case must be 3.

Input

Each test case is a number between 1 and 10^240.