Add : Codechef (Dec challenge) WA
You are given an array A of size N. Let us list down all the subarrays of the given array. There will be a total of N * (N + 1) / 2 subarrays of the given array. Let us sort each of the subarrays in descending order of the numbers in it.

Now you want to sort these subarrays in descending order. You can compare two subarrays B, C, as follows.


    compare(B, C):
        Append N - |B| zeros at the end of the array B.
        Append N - |C| zeros at the end of the array C.
        for i = 1 to N:
            if B[i] < C[i]:
                return B is less than C
            if B[i] > C[i]:
                return B is greater than C
        return B and C are equal.

You are given M queries asking for the maximum element in the pth subarray (1-based indexing).