Add : Hackerrank (Sintaxis-Experto)
Varun is mad about cryptography, that is why he writes Vernota encoded messages. (Vernota is encoding technique used by Varun)

He calls the lucky letter in a word the letter which is in the middle of the word. If the word's length is even, the lucky letter is the left of the two middle letters.

In the following examples, the lucky letter is highlighted: ra h ul, s a ra . If the word consists of single letter, then according to above definition this letter is the lucky letter.

Varun encodes each word in the following way: he writes down the lucky letter of the word, then deletes it and repeats the process until there are no letters left.

For example, he encodes the word varun as rauvn.

You are given encoded word.Can you decode it?