Add : Hackerrank (sintaxis-experto)
Every year PACE organizes it's mega-event named Prudence.

This time PACE is going to conduct a special sub-event named XYZ. It is going to be really fun.

There are n spots in a straight path. Each spot has its own unique ID from 1 to n. Participant will be given starting spot a and ending spot b.

PACE is providing a special flying suit to participants.

image

To get to the spot b, participant can fly using the flying suit between any pair of spots any number of times, but he/she has to start his/her journey at the spot a and finish it at the spot b.

Each spot belongs to one of two categories.The cost of flight from the spot i to the spot j is zero if both spots belong to the same category, and |i - j| if they belong to different categories.

The participant who will complete his/her journey at minimum cost will be the winner.

Print the minimum cost participant has to pay to get to the destination.