Add : Codechef (Codevaitam 3.0)
Chef's father gave him a integer N and asked him to count all such pair of x and y such that xy is equal to N i.e., x raised to the power y is equal to N. Chef also wants to calculate summation of (x+y) for all the pairs satisfying the above condition. Help Chef in solving this problem.