Add : Codechef (Feb 17 Challenge) subtask 1
Chef is long been involved in cooking recipes. He has now got bored with it, and has to decided to play with food instead of cooking it up.

He has already selected N spaghetti strands for his dish. He has to select just one extra strand such that this strand's length makes a non-degenerate triangle with some of the two other strands already present in the dish. A triangle formed by three collinear points is called a degenerate triangle. The length of the strand to be selected must be an integer in the range [L, R] (both inclusive).

Formally, let array L denote the lengths of N strands, then the extra strand of length X can be selected if there exists two different indices i, j (i ≠ j) such that there can be made a triangle of side lengths Li, Lj and X.

Can you help him find the number of possible valid lengths of the extra strand?

Note again that three collinear points are not considered to form a triangle.