Add : Codechef (March challenge)
A sequence is called nice if its elements are distinct consecutive numbers, possibly in changed order. For example, both (6, 7, 8) and (15, 13, 16, 14) are nice, while (4, 6), (4, 5, 5, 6) and (15, 16, 15) are not.

Limak has a nice sequence. While he was in school today, someone inserted one extra number in the sequence. Limak has just returned and realized that the sequence isn't nice anymore! He wants to remove the inserted number and make his sequence nice again. Can you help him to find the number that he should remove?

Formally, in each test case you are given a sequence of N numbers A1, A2, ..., AN. Your task is to find the value X, such that removing one occurrence of X would make the sequence nice. It's guaranteed that exactly one solution exists.