Add : Codechef (Cook 81)
Limak has a string S, that consists of N lowercase English letters. Limak then created a new string by repeating S exactly K times. For example, for S = "abcb" and K = 2, he would get "abcbabcb".

Your task is to count the number of subsequences "ab" (not necessarily consecutive) in the new string.

In other words, find the number pairs of indices i < j, such that the i-th and j-th characters in the new string are 'a' and 'b' respectively.